requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
crispy-bootstrap5==0.7
Django==4.1.3
django-crispy-forms==1.14.0
idna==3.4
requests==2.28.1
sqlparse==0.4.3
tmdbsimple==2.9.1
tzdata==2022.7
urllib3==1.26.13
 
python-dotenv
whitenoise
